General introduction

Installation

Installing the package follows the typical Julia scheme

julia> ]
pkg> add BEAVARs
pkg>

and, after pressing backspace to get back to the julia terminal, typing

julia> using BEAVARs

Usage

The main function of the package is

beavar(model_type, set_strct, hyp_strct, data_strct)

which calls the relevant models and performs the estimation. Using the package boils down to the correct specification of these arguments, for which special functions exist to help you create them. Before jumping in the details let's give a brief overview.

model_type: An object that allows Julia to know which model you want to use and call the relevant functions.

set_strct: A structure containing the general VAR setup such as number of lags, number of draws, etc.

hyp_strct: A structure for setting the hyperparameters for the Bayesian estimation.

data_strct: A structure containing your data.

The first three objects are generated using a helper function make_setup.

BEAVARs.makeSetup — Method

model_type, set_strct, hyp_strct = makeSetup(model_str::String; p::Int=4,n_burn::Int=1000,n_save::Int=1000,n_irf::Int=16,n_fcst::Int = 8,hyp::BVARmodelHypSetup=hypDefault_strct())

Specify a model and generate structures for the Bayesian VAR and the hyperparameters.

Only the first argument is mandatory, rest is optional with default values.

Arguments

model_str: String, currently supported are "CPZ2023", "Chan2020minn", "Chan2020csv", "Chan2020iniw", "BGR2010"
p:         number of lags, default is 4
n_burn:    number of burn-in draws that will be discarded, default is 2000
n_save:    number of retained draws (total is then nburn + nsave), default is 1000
n_irf:     horizon of impulse responses, default is 16
n_fcst:    horizon of forecasting periods, default is 8
hyp:       hyperparameter structure populated with default values for each model. See the relevant papers/documentation for details. To generate your own see the relevant structures below.

See also hypChan2020, hypBGR2010.

The last is generated using the helper function

makeDataSetup(model_type, data_tab::TimeArray; var_list::Vector{Symbol})

which is different for evey model.

Next we will estimate a model using a particular example.